refactor to use visitor pattern
