+++ src/docgen/view/enum-member.ts
@@ -1,50 +1,9 @@
 import * as reflect from 'jsii-reflect';
-import { defaultAnchorFormatter, MarkdownDocument } from '../render/markdown-doc';
 import { EnumMemberSchema } from '../schema';
 import { Transpile, TranspiledEnumMember } from '../transpile/transpile';
 import { extractDocs } from '../util';
-import { MarkdownRenderContext } from './documentation';
 
 export class EnumMember {
-  public static toMarkdown(
-    em: EnumMemberSchema,
-    context: MarkdownRenderContext,
-  ): MarkdownDocument {
-    const anchorFormatter = context.anchorFormatter ?? defaultAnchorFormatter;
-
-    const md = new MarkdownDocument({
-      id: anchorFormatter({
-        id: em.id,
-        displayName: em.displayName,
-        fqn: em.fqn,
-        packageName: context.packageName,
-        packageVersion: context.packageVersion,
-        submodule: context.submodule,
-      }),
-      header: {
-        title: em.fqn.split('.').pop(),
-        pre: true,
-        strike: em.docs.deprecated,
-      },
-    });
-
-    if (em.docs.deprecated) {
-      md.bullet(
-        `${MarkdownDocument.italic('Deprecated:')} ${em.docs.deprecationReason}`,
-      );
-      md.lines('');
-    }
-
-    if (em.docs) {
-      md.docs(em.docs);
-    }
-
-    md.split();
-    md.lines('');
-
-    return md;
-  }
-
   private readonly transpiled: TranspiledEnumMember;
   constructor(transpile: Transpile, private readonly em: reflect.EnumMember) {
     this.transpiled = transpile.enumMember(em);
