+++ src/docgen/view/interface.ts
@@ -1,59 +1,11 @@
 import * as reflect from 'jsii-reflect';
-import { defaultAnchorFormatter, defaultLinkFormatter, MarkdownDocument } from '../render/markdown-doc';
 import { InterfaceSchema, JsiiEntity } from '../schema';
 import { Transpile, TranspiledInterface } from '../transpile/transpile';
 import { extractDocs } from '../util';
-import { MarkdownRenderContext } from './documentation';
 import { InstanceMethods } from './instance-methods';
 import { Properties } from './properties';
 
 export class Interface {
-  public static toMarkdown(
-    iface: InterfaceSchema,
-    context: MarkdownRenderContext,
-  ): MarkdownDocument {
-    const anchorFormatter = context.anchorFormatter ?? defaultAnchorFormatter;
-    const linkFormatter = context.linkFormatter ?? defaultLinkFormatter;
-
-    const md = new MarkdownDocument({
-      id: anchorFormatter({
-        id: iface.id,
-        displayName: iface.displayName,
-        fqn: iface.fqn,
-        packageName: context.packageName,
-        packageVersion: context.packageVersion,
-        submodule: context.submodule,
-      }),
-      header: { title: iface.fqn.split('.').pop() },
-    });
-
-    if (iface.interfaces.length > 0) {
-      const bases = [];
-      for (const base of iface.interfaces) {
-        bases.push(linkFormatter(base));
-      }
-      md.bullet(`${MarkdownDocument.italic('Extends:')} ${bases.join(', ')}`);
-      md.lines('');
-    }
-
-    if (iface.implementations.length > 0) {
-      const impls = [];
-      for (const impl of iface.implementations) {
-        impls.push(linkFormatter(impl));
-      }
-      md.bullet(`${MarkdownDocument.italic('Implemented By:')} ${impls.join(', ')}`);
-      md.lines('');
-    }
-
-    if (iface.docs) {
-      md.docs(iface.docs);
-    }
-
-    md.section(InstanceMethods.toMarkdown(iface.instanceMethods, context));
-    md.section(Properties.toMarkdown(iface.properties, context));
-    return md;
-  }
-
   public static isStruct(iface: reflect.InterfaceType): boolean {
     return iface.datatype;
   }
