+++ src/docgen/view/static-function.ts
@@ -1,48 +1,10 @@
 import * as reflect from 'jsii-reflect';
-import { defaultAnchorFormatter, MarkdownDocument } from '../render/markdown-doc';
 import { MethodSchema } from '../schema';
 import { Transpile, TranspiledCallable } from '../transpile/transpile';
 import { extractDocs } from '../util';
-import { MarkdownRenderContext } from './documentation';
 import { Parameter } from './parameter';
 
 export class StaticFunction {
-  public static toMarkdown(
-    method: MethodSchema,
-    context: MarkdownRenderContext,
-  ) {
-    const anchorFormatter = context.anchorFormatter ?? defaultAnchorFormatter;
-
-    const md = new MarkdownDocument({
-      id: anchorFormatter({
-        id: method.id,
-        displayName: method.displayName,
-        fqn: method.fqn,
-        packageName: context.packageName,
-        packageVersion: context.packageVersion,
-        submodule: context.submodule,
-      }),
-      header: {
-        title: method.fqn.split('.').pop(),
-        pre: true,
-        strike: method.docs.deprecated,
-      },
-    });
-
-    if (method.usage) {
-      md.code(
-        context.language.toString(),
-        method.usage,
-      );
-    }
-
-    for (const param of method.parameters) {
-      md.section(Parameter.toMarkdown(param, context));
-    }
-
-    return md;
-  }
-
   private readonly transpiled: TranspiledCallable;
   private readonly parameters: Parameter[];
   constructor(
