+++ src/docgen/view/struct.ts
@@ -1,51 +1,10 @@
 import * as reflect from 'jsii-reflect';
-import { defaultAnchorFormatter, MarkdownDocument } from '../render/markdown-doc';
 import { StructSchema } from '../schema';
 import { Transpile, TranspiledStruct } from '../transpile/transpile';
 import { extractDocs } from '../util';
-import { MarkdownRenderContext } from './documentation';
 import { Properties } from './properties';
 
 export class Struct {
-  public static toMarkdown(
-    struct: StructSchema,
-    context: MarkdownRenderContext,
-  ): MarkdownDocument {
-    const anchorFormatter = context.anchorFormatter ?? defaultAnchorFormatter;
-
-    const md = new MarkdownDocument({
-      id: anchorFormatter({
-        id: struct.id,
-        displayName: struct.displayName,
-        fqn: struct.fqn,
-        packageName: context.packageName,
-        packageVersion: context.packageVersion,
-        submodule: context.submodule,
-      }),
-      header: { title: struct.fqn.split('.').pop() },
-    });
-
-    if (struct.docs) {
-      md.docs(struct.docs);
-    }
-
-    const initializer = new MarkdownDocument({
-      id: `${struct.id}.Initializer`,
-      header: { title: 'Initializer' },
-    });
-
-    if (struct.usage) {
-      initializer.code(
-        context.language.toString(),
-        struct.usage,
-      );
-    }
-
-    md.section(initializer);
-    md.section(Properties.toMarkdown(struct.properties, context));
-    return md;
-  }
-
   private readonly transpiled: TranspiledStruct;
   private readonly properties: Properties;
   constructor(
