+++ src/index.ts
@@ -1,5 +1,5 @@
-export { Documentation } from './docgen/view/documentation';
+export { Documentation, getTranspilerForLanguage } from './docgen/view/documentation';
 export { Language } from './docgen/transpile/transpile';
 export { UnsupportedLanguageError } from './docgen/transpile/transpile';
-export { TranspiledType } from './docgen/transpile/transpile';
 export * from './errors';
+export * from './docgen/schema';
