+++ test/docgen/view/class.test.ts
@@ -1,10 +1,8 @@
 import * as reflect from 'jsii-reflect';
-import { Language } from '../../../lib';
-import { CSharpTranspile } from '../../../src/docgen/transpile/csharp';
-import { JavaTranspile } from '../../../src/docgen/transpile/java';
-import { PythonTranspile } from '../../../src/docgen/transpile/python';
-import { TypeScriptTranspile } from '../../../src/docgen/transpile/typescript';
+import { Language } from '../../../src';
+import { MarkdownRenderer } from '../../../src/docgen/render/markdown-render';
 import { Class } from '../../../src/docgen/view/class';
+import { getTranspilerForLanguage } from '../../../src/docgen/view/documentation';
 import { Assemblies } from '../assemblies';
 
 const assembly: reflect.Assembly = Assemblies.instance.withoutSubmodules;
@@ -21,42 +19,10 @@ const findClass = (): reflect.ClassType => {
   throw new Error('Assembly does not contain a class');
 };
 
-describe('python', () => {
-  const transpile = new PythonTranspile();
-  test('snapshot', () => {
-    const klass = new Class(transpile, findClass()).toJson();
-    const markdown = Class.toMarkdown(klass, { language: Language.PYTHON, ...metadata });
-    expect(klass).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('typescript', () => {
-  const transpile = new TypeScriptTranspile();
-  test('snapshot', () => {
-    const klass = new Class(transpile, findClass()).toJson();
-    const markdown = Class.toMarkdown(klass, { language: Language.TYPESCRIPT, ...metadata });
-    expect(klass).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('java', () => {
-  const transpile = new JavaTranspile();
-  test('snapshot', () => {
-    const klass = new Class(transpile, findClass()).toJson();
-    const markdown = Class.toMarkdown(klass, { language: Language.JAVA, ...metadata });
-    expect(klass).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('csharp', () => {
-  const transpile = new CSharpTranspile();
-  test('snapshot', () => {
-    const klass = new Class(transpile, findClass()).toJson();
-    const markdown = Class.toMarkdown(klass, { language: Language.CSHARP, ...metadata });
-    expect(klass).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
+test.each(Language.values())('%s snapshot', (language) => {
+  const transpile = getTranspilerForLanguage(language);
+  const markdown = new MarkdownRenderer({ language, ...metadata });
+  const klass = new Class(transpile, findClass()).toJson();
+  expect(klass).toMatchSnapshot();
+  expect(markdown.visitClass(klass).render()).toMatchSnapshot();
 });
