+++ test/docgen/view/initializer.test.ts
@@ -1,9 +1,7 @@
 import * as reflect from 'jsii-reflect';
-import { Language } from '../../../lib';
-import { CSharpTranspile } from '../../../src/docgen/transpile/csharp';
-import { JavaTranspile } from '../../../src/docgen/transpile/java';
-import { PythonTranspile } from '../../../src/docgen/transpile/python';
-import { TypeScriptTranspile } from '../../../src/docgen/transpile/typescript';
+import { Language } from '../../../src';
+import { MarkdownRenderer } from '../../../src/docgen/render/markdown-render';
+import { getTranspilerForLanguage } from '../../../src/docgen/view/documentation';
 import { Initializer } from '../../../src/docgen/view/initializer';
 import { Assemblies } from '../assemblies';
 
@@ -23,42 +21,10 @@ const findInitializer = (): reflect.Initializer => {
   throw new Error('Assembly does not contain an initializer');
 };
 
-describe('python', () => {
-  const transpile = new PythonTranspile();
-  test('snapshot', () => {
-    const initializer = new Initializer(transpile, findInitializer()).toJson();
-    const markdown = Initializer.toMarkdown(initializer, { language: Language.PYTHON, ...metadata });
-    expect(initializer).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('typescript', () => {
-  const transpile = new TypeScriptTranspile();
-  test('snapshot', () => {
-    const initializer = new Initializer(transpile, findInitializer()).toJson();
-    const markdown = Initializer.toMarkdown(initializer, { language: Language.TYPESCRIPT, ...metadata });
-    expect(initializer).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('java', () => {
-  const transpile = new JavaTranspile();
-  test('snapshot', () => {
-    const initializer = new Initializer(transpile, findInitializer()).toJson();
-    const markdown = Initializer.toMarkdown(initializer, { language: Language.JAVA, ...metadata });
-    expect(initializer).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('csharp', () => {
-  const transpile = new CSharpTranspile();
-  test('snapshot', () => {
-    const initializer = new Initializer(transpile, findInitializer()).toJson();
-    const markdown = Initializer.toMarkdown(initializer, { language: Language.CSHARP, ...metadata });
-    expect(initializer).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
+test.each(Language.values())('%s snapshot', (language) => {
+  const transpile = getTranspilerForLanguage(language);
+  const markdown = new MarkdownRenderer({ language, ...metadata });
+  const init = new Initializer(transpile, findInitializer()).toJson();
+  expect(init).toMatchSnapshot();
+  expect(markdown.visitInitializer(init).render()).toMatchSnapshot();
 });
