+++ test/docgen/view/instance-method.test.ts
@@ -1,9 +1,7 @@
 import * as reflect from 'jsii-reflect';
-import { CSharpTranspile } from '../../../src/docgen/transpile/csharp';
-import { JavaTranspile } from '../../../src/docgen/transpile/java';
-import { PythonTranspile } from '../../../src/docgen/transpile/python';
+import { MarkdownRenderer } from '../../../src/docgen/render/markdown-render';
 import { Language } from '../../../src/docgen/transpile/transpile';
-import { TypeScriptTranspile } from '../../../src/docgen/transpile/typescript';
+import { getTranspilerForLanguage } from '../../../src/docgen/view/documentation';
 import { InstanceMethod } from '../../../src/docgen/view/instance-method';
 import { Assemblies } from '../assemblies';
 
@@ -25,42 +23,10 @@ const findInstanceMethod = (): reflect.Method => {
   throw new Error('Assembly does not contain an instance method');
 };
 
-describe('python', () => {
-  const transpile = new PythonTranspile();
-  test('snapshot', () => {
-    const instanceMethod = new InstanceMethod(transpile, findInstanceMethod()).toJson();
-    const markdown = InstanceMethod.toMarkdown(instanceMethod, { language: Language.PYTHON, ...metadata });
-    expect(instanceMethod).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('typescript', () => {
-  const transpile = new TypeScriptTranspile();
-  test('snapshot', () => {
-    const instanceMethod = new InstanceMethod(transpile, findInstanceMethod()).toJson();
-    const markdown = InstanceMethod.toMarkdown(instanceMethod, { language: Language.TYPESCRIPT, ...metadata });
-    expect(instanceMethod).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('java', () => {
-  const transpile = new JavaTranspile();
-  test('snapshot', () => {
-    const instanceMethod = new InstanceMethod(transpile, findInstanceMethod()).toJson();
-    const markdown = InstanceMethod.toMarkdown(instanceMethod, { language: Language.JAVA, ...metadata });
-    expect(instanceMethod).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('csharp', () => {
-  const transpile = new CSharpTranspile();
-  test('snapshot', () => {
-    const instanceMethod = new InstanceMethod(transpile, findInstanceMethod()).toJson();
-    const markdown = InstanceMethod.toMarkdown(instanceMethod, { language: Language.CSHARP, ...metadata });
-    expect(instanceMethod).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
+test.each(Language.values())('%s snapshot', (language) => {
+  const transpile = getTranspilerForLanguage(language);
+  const markdown = new MarkdownRenderer({ language, ...metadata });
+  const method = new InstanceMethod(transpile, findInstanceMethod()).toJson();
+  expect(method).toMatchSnapshot();
+  expect(markdown.visitInstanceMethod(method).render()).toMatchSnapshot();
 });
