+++ test/docgen/view/interface.test.ts
@@ -1,9 +1,7 @@
 import * as reflect from 'jsii-reflect';
-import { CSharpTranspile } from '../../../src/docgen/transpile/csharp';
-import { JavaTranspile } from '../../../src/docgen/transpile/java';
-import { PythonTranspile } from '../../../src/docgen/transpile/python';
+import { MarkdownRenderer } from '../../../src/docgen/render/markdown-render';
 import { Language } from '../../../src/docgen/transpile/transpile';
-import { TypeScriptTranspile } from '../../../src/docgen/transpile/typescript';
+import { getTranspilerForLanguage } from '../../../src/docgen/view/documentation';
 import { Interface } from '../../../src/docgen/view/interface';
 import { Assemblies } from '../assemblies';
 
@@ -23,42 +21,10 @@ const findInterface = (): reflect.InterfaceType => {
   throw new Error('Assembly does not contain an interface');
 };
 
-describe('python', () => {
-  const transpile = new PythonTranspile();
-  test('snapshot', () => {
-    const iface = new Interface(transpile, findInterface()).toJson();
-    const markdown = Interface.toMarkdown(iface, { language: Language.PYTHON, ...metadata });
-    expect(iface).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('typescript', () => {
-  const transpile = new TypeScriptTranspile();
-  test('snapshot', () => {
-    const iface = new Interface(transpile, findInterface()).toJson();
-    const markdown = Interface.toMarkdown(iface, { language: Language.TYPESCRIPT, ...metadata });
-    expect(iface).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('java', () => {
-  const transpile = new JavaTranspile();
-  test('snapshot', () => {
-    const iface = new Interface(transpile, findInterface()).toJson();
-    const markdown = Interface.toMarkdown(iface, { language: Language.JAVA, ...metadata });
-    expect(iface).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('csharp', () => {
-  const transpile = new CSharpTranspile();
-  test('snapshot', () => {
-    const iface = new Interface(transpile, findInterface()).toJson();
-    const markdown = Interface.toMarkdown(iface, { language: Language.CSHARP, ...metadata });
-    expect(iface).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
+test.each(Language.values())('%s snapshot', (language) => {
+  const transpile = getTranspilerForLanguage(language);
+  const markdown = new MarkdownRenderer({ language, ...metadata });
+  const iface = new Interface(transpile, findInterface()).toJson();
+  expect(iface).toMatchSnapshot();
+  expect(markdown.visitInterface(iface).render()).toMatchSnapshot();
 });
