+++ test/docgen/view/property.test.ts
@@ -1,9 +1,7 @@
 import * as reflect from 'jsii-reflect';
-import { Language } from '../../../lib';
-import { CSharpTranspile } from '../../../src/docgen/transpile/csharp';
-import { JavaTranspile } from '../../../src/docgen/transpile/java';
-import { PythonTranspile } from '../../../src/docgen/transpile/python';
-import { TypeScriptTranspile } from '../../../src/docgen/transpile/typescript';
+import { Language } from '../../../src';
+import { MarkdownRenderer } from '../../../src/docgen/render/markdown-render';
+import { getTranspilerForLanguage } from '../../../src/docgen/view/documentation';
 import { Property } from '../../../src/docgen/view/property';
 import { Assemblies } from '../assemblies';
 
@@ -23,42 +21,10 @@ const findProperty = (): reflect.Property => {
   throw new Error('Assembly does not contain a property');
 };
 
-describe('python', () => {
-  const transpile = new PythonTranspile();
-  test('snapshot', () => {
-    const parameter = new Property(transpile, findProperty()).toJson();
-    const markdown = Property.toMarkdown(parameter, { language: Language.PYTHON, ...metadata });
-    expect(parameter).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('typescript', () => {
-  const transpile = new TypeScriptTranspile();
-  test('snapshot', () => {
-    const parameter = new Property(transpile, findProperty()).toJson();
-    const markdown = Property.toMarkdown(parameter, { language: Language.TYPESCRIPT, ...metadata });
-    expect(parameter).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('java', () => {
-  const transpile = new JavaTranspile();
-  test('snapshot', () => {
-    const parameter = new Property(transpile, findProperty()).toJson();
-    const markdown = Property.toMarkdown(parameter, { language: Language.JAVA, ...metadata });
-    expect(parameter).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
-});
-
-describe('csharp', () => {
-  const transpile = new CSharpTranspile();
-  test('snapshot', () => {
-    const parameter = new Property(transpile, findProperty()).toJson();
-    const markdown = Property.toMarkdown(parameter, { language: Language.CSHARP, ...metadata });
-    expect(parameter).toMatchSnapshot();
-    expect(markdown.render()).toMatchSnapshot();
-  });
+test.each(Language.values())('%s snapshot', (language) => {
+  const transpile = getTranspilerForLanguage(language);
+  const markdown = new MarkdownRenderer({ language, ...metadata });
+  const prop = new Property(transpile, findProperty()).toJson();
+  expect(prop).toMatchSnapshot();
+  expect(markdown.visitProperty(prop).render()).toMatchSnapshot();
 });
